Create a class for Date converting to timestamp
